Create stock.py
This program used pandas_datareader to get data from yahoo finance. It plots the historical adjusted daily returns over a specified time period.

A histogram of the daily returns over this time is also plotted with an overlay of a normal distribution.

To check how well this distribution fits a normal a QQ-plot is generated.