Add future price prediction & investment scenario
There has been an update to use the data to predict the future price of the company shares. This can be done at any time ahead of the latest available price. The model of the price used is a normal distribution. The confidence in the prediction is set at 95% (X=2).

The investment scenario involves imagining a principal has been invested of £1M on the first day of the range considered. The output tells us how our investment could have done.

A comparison of this investment scenario will come soon where a £1M savings price will be considered over the same time period.