volatility and drift calculation changes
There has been changes to the calculation of the volatility and drift. This comes from calculating returns for a normal and a lognormal model.

The new way of getting return data required that in the sm.qqplot() function, the list of data had to be converted to an array.

I added a T variable at the beginning to be able to easily edit the length of the financial day.